:recycle: Seperate deposit logic to utils
