+++ backend/src/utils/index.js
@@ -1,7 +1,7 @@
 export { send } from './wallet.js';
 export { registerAddress } from './register.js';
-export { newDeposit } from './deposit.js';
+export { newDeposit, depositUser } from './deposit.js';
 export { newSession } from './session.js';
 export { default as api } from './api.js';
-export * from './utils.js';
-export * from './verify.js';
+export { formatCoin, changeAddressPrefix, validateAddress } from './utils.js';
+export { verifyUser, getBalance } from './verify.js';
