+++ package.json
@@ -14,6 +14,7 @@
     "start": "yarn --cwd backend start",
     "dev": "NODE_ENV=development yarn --cwd backend dev & yarn --cwd frontend dev",
     "lint": "yarn --cwd backend lint & yarn --cwd frontend lint",
-    "generate": "yarn --cwd frontend generate"
+    "generate": "yarn --cwd frontend generate",
+    "delete": "yarn --cwd backend delete"
   }
 }
